+++ BTDBTest/SerializationTests/BonSerializerTest.cs
@@ -40,6 +40,19 @@ public class BonSerializerTest
         public Dictionary<long, int>? LongIntDict;
     }
 
+    [Generate]
+    public class TupleArityTypes
+    {
+        public (int A, string B, long C) ValueTuple3;
+        public (byte A, long B, string C, int D) ValueTuple4;
+        public (string A, byte B, double C, int D, short E) ValueTuple5;
+        public (DateTime A, Guid B, string C) ValueTupleDateTimeGuid;
+        public Tuple<int, string, long>? Tuple3;
+        public Tuple<byte, long, string, int>? Tuple4;
+        public Tuple<string, byte, double, int, short>? Tuple5;
+        public Tuple<DateTime, Guid, string>? TupleDateTimeGuid;
+    }
+
     [Fact]
     public void SerializeDeserializeAllSupportedTypes()
     {
@@ -82,6 +95,53 @@ public class BonSerializerTest
         var (a, b) = (Tuple<object, object>)deserialized;
         Assert.Equal(42u, (ulong)a);
         Assert.Equal(424242424242L, (long)(ulong)b);
+    }
+
+    [Fact]
+    public void SerializeDeserializeTuplesWithMoreThanTwoItems()
+    {
+        var obj = new TupleArityTypes
+        {
+            ValueTuple3 = (11, "two", 333L),
+            ValueTuple4 = (1, 22L, "three", 4444),
+            ValueTuple5 = ("one", 2, 3.5, 4, 5),
+            ValueTupleDateTimeGuid = (new(2025, 5, 25, 14, 30, 0),
+                Guid.Parse("d1f4fd89-f6cd-427e-a703-61c60ea46f76"), "dt-guid"),
+            Tuple3 = new(11, "two", 333L),
+            Tuple4 = new(1, 22L, "three", 4444),
+            Tuple5 = new("one", 2, 3.5, 4, 5),
+            TupleDateTimeGuid = new(new(2026, 1, 2, 3, 4, 5),
+                Guid.Parse("364b9a4f-207a-44d1-aebb-3bbffb24e5c8"), "tuple-dt-guid")
+        };
+        var builder = new BonBuilder();
+        BonSerializerFactory.Serialize(ref builder, obj);
+        var bon = new Bon(builder.FinishAsMemory());
+        var deserialized = Assert.IsType<TupleArityTypes>(BonSerializerFactory.Deserialize(ref bon));
+
+        Assert.Equal(obj.ValueTuple3, deserialized.ValueTuple3);
+        Assert.Equal(obj.ValueTuple4, deserialized.ValueTuple4);
+        Assert.Equal(obj.ValueTuple5, deserialized.ValueTuple5);
+        Assert.Equal(obj.ValueTupleDateTimeGuid, deserialized.ValueTupleDateTimeGuid);
+        Assert.Equal(obj.Tuple3, deserialized.Tuple3);
+        Assert.Equal(obj.Tuple4, deserialized.Tuple4);
+        Assert.Equal(obj.Tuple5, deserialized.Tuple5);
+        Assert.Equal(obj.TupleDateTimeGuid, deserialized.TupleDateTimeGuid);
+    }
+
+    [Fact]
+    public void DeserializeBoxedTupleWithMoreThanTwoItems()
+    {
+        var obj = (object)(11, "two", 333L, 4.5);
+        var builder = new BonBuilder();
+        BonSerializerFactory.Serialize(ref builder, obj);
+        var bon = new Bon(builder.FinishAsMemory());
+        var deserialized =
+            Assert.IsType<Tuple<object?, object?, object?, object?>>(BonSerializerFactory.Deserialize(ref bon));
+
+        Assert.Equal(11L, Convert.ToInt64(deserialized.Item1));
+        Assert.Equal("two", deserialized.Item2);
+        Assert.Equal(333L, Convert.ToInt64(deserialized.Item3));
+        Assert.Equal(4.5, deserialized.Item4);
     }
 
     [Fact]